--- wws/src/percent.rs
@@ -1,62 +1,0 @@
-/*
- * percent.rs
- *
- * Wilson's Web Server - Serves a zoo of user-generated content
- * Copyright (C) 2019-2026 Wikijump Team
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program. If not, see <http://www.gnu.org/licenses/>.
- */
-
-use std::fmt::Write;
-
-pub fn percent_encode_rfc5987(s: &str) -> String {
-    let mut out = String::with_capacity(s.len());
-    for byte in s.as_bytes() {
-        match byte {
-            b'A'..=b'Z'
-            | b'a'..=b'z'
-            | b'0'..=b'9'
-            | b'!'
-            | b'#'
-            | b'$'
-            | b'&'
-            | b'+'
-            | b'-'
-            | b'.'
-            | b'^'
-            | b'_'
-            | b'`'
-            | b'|'
-            | b'~' => {
-                out.push(char::from(*byte));
-            }
-            _ => {
-                let _ = write!(out, "%{byte:02X}");
-            }
-        }
-    }
-    out
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn rfc5987_encode() {
-        assert_eq!(percent_encode_rfc5987("hello world"), "hello%20world");
-        assert_eq!(percent_encode_rfc5987("a/b"), "a%2Fb");
-        assert_eq!(percent_encode_rfc5987("simple"), "simple");
-    }
-}
